Update Hostgroup client and add stratos validation (#361)
* Update Hostgroup client and add stratos validation

* Update name to hostGroup pattern

* Rename file
